requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai==1.57.0
pydantic>=2.9.0
pillow>=10.1.0
python-dotenv==1.0.0
